+++ investigraph/cache.py
@@ -1,15 +1,15 @@
+import logging
 from functools import cache
 from typing import Any, Iterable, Set
 
 import fakeredis
 import redis
-import shortuuid
 from cachelib.serializers import RedisSerializer
 
 from investigraph import settings
-from investigraph.logging import get_logger
+from investigraph.util import data_checksum
 
-log = get_logger(__name__)
+log = logging.getLogger(__name__)
 
 
 class Cache:
@@ -17,7 +17,7 @@ class Cache:
     This is an extremely simple cache interface for sharing tasks data
     efficiently via redis (or fakeredis during development)
 
-    it creates (prefixed) random keys during data set to cache.
+    it creates (prefixed) keys based on input data
 
     it mimics redis GETDEL so that after fetching data from cache the key is
     deleted (turn of by `delete=False`)
@@ -28,19 +28,21 @@ class Cache:
     def __init__(self):
         if settings.DEBUG:
             con = fakeredis.FakeStrictRedis()
+            con.ping()
+            log.info("Redis connected: `fakeredis`")
         else:
             con = redis.from_url(settings.REDIS_URL)
-        con.ping()
-        log.info("Redis initialized", url=settings.REDIS_URL)
+            con.ping()
+            log.info("Redis connected: `{settings.REDIS_URL}`")
         self.cache = con
 
     def set(self, data: Any, key: str | None = None) -> str:
+        key = key or data_checksum(data)
         data = self.serializer.dumps(data)
-        key = key or shortuuid.uuid()
         self.cache.set(self.get_key(key), data)
         return key
 
-    def get(self, key: str, delete: bool | None = True) -> Any:
+    def get(self, key: str, delete: bool | None = False) -> Any:
         key = self.get_key(key)
         res = self.cache.get(key)
         if delete:
@@ -50,12 +52,12 @@ class Cache:
             return data
 
     def sadd(self, *values: Iterable[Any], key: str | None = None) -> str:
-        key = key or shortuuid.uuid()
         values = [str(v) for v in values]
+        key = key or data_checksum(values)
         self.cache.sadd(self.get_key(key) + "#SET", *values)
         return key
 
-    def smembers(self, key: str, delete: bool | None = True) -> Set[str]:
+    def smembers(self, key: str, delete: bool | None = False) -> Set[str]:
         key = self.get_key(key) + "#SET"
         res: Set[bytes] = self.cache.smembers(key)
         if delete:
@@ -67,7 +69,7 @@ class Cache:
 
     @staticmethod
     def get_key(key: str) -> str:
-        return f"{settings.CACHE_PREFIX}:{key}"
+        return f"{settings.REDIS_PREFIX}:{key}"
 
 
 @cache