+++ investigraph/logic/extract.py
@@ -2,8 +2,10 @@
 Extract sources to iterate objects to dict records
 """
 
+import numpy as np
 import pandas as pd
 from pantomime import types
+from runpandarun.io import guess_handler_from_mimetype
 
 from investigraph.model.context import Context
 from investigraph.model.resolver import Resolver
@@ -15,8 +17,7 @@ JSON = types.JSON
 
 def yield_pandas(df: pd.DataFrame) -> RecordGenerator:
     for _, row in df.iterrows():
-        row = {k: v if not pd.isna(v) else None for k, v in row.items()}
-        yield row
+        yield dict(row.replace(np.nan, None))
 
 
 def extract_pandas(
@@ -24,6 +25,8 @@ def extract_pandas(
 ) -> RecordGenerator:
     play = resolver.source.pandas
     play.read.uri = resolver.source.uri
+    if play.read.handler is None:
+        play.read.handler = f"read_{guess_handler_from_mimetype(resolver.mimetype)}"
     for ix, chunk in enumerate(resolver.iter(chunk_size)):
         df = play.read.handle(chunk)
         if resolver.mimetype == types.CSV: