+++ investigraph/settings.py
@@ -1,15 +1,17 @@
 import os
+from datetime import datetime, timedelta
 from pathlib import Path
 from typing import Any
 
 from banal import as_bool
 
-from investigraph import __version__
-
 
 def get_env(env: str, default: Any | None = None) -> Any | None:
     return os.environ.get(env, default)
 
+
+VERSION = "0.3.0"
+RUN_TIME = datetime.utcnow().replace(microsecond=0)
 
 DEBUG = as_bool(get_env("DEBUG", 1))
 DATA_ROOT = Path(get_env("DATA_ROOT", Path.cwd() / "data")).absolute()
@@ -23,10 +25,15 @@ DEFAULT_TRANSFORMER = get_env(
 DEFAULT_LOADER = get_env("DEFAULT_LOADER", "investigraph.logic.load:load_proxies")
 
 REDIS_URL = get_env("REDIS_URL", "redis://localhost:6379")
-CACHE_PREFIX = get_env("CACHE_PREFIX", f"investigraph:{__version__}")
+REDIS_PREFIX = get_env("REDIS_PREFIX", f"investigraph:{VERSION}")
 
-FETCH_RETRIES = int(get_env("FETCH_RETRIES", 3))
-FETCH_RETRY_DELAY = int(get_env("FETCH_RETRY_DELAY", 5))
+TASK_CACHE = as_bool(get_env("TASK_CACHE", 1))
+TASK_RETRIES = int(get_env("TASK_RETRIES", 3))
+TASK_RETRY_DELAY = int(get_env("TASK_RETRY_DELAY", 5))
+TASK_CACHE_EXPIRATION = int(get_env("TASK_CACHE_EXPIRATION", 0)) or None  # in minutes
+TASK_CACHE_EXPIRATION = (
+    timedelta(TASK_CACHE_EXPIRATION) if TASK_CACHE_EXPIRATION is not None else None
+)
 
 TASK_RUNNER = get_env("PREFECT_TASK_RUNNER", "").lower()