+++ tests/test_pipeline.py
@@ -1,12 +1,14 @@
-# import os
+from importlib import reload
 
+import cloudpickle
 from ftmq.io import smart_read_proxies
 from ftmstore import get_dataset
 from moto import mock_s3
 
+from investigraph import settings
 from investigraph.model import DatasetBlock, FlowOptions
+from investigraph.model.context import init_context
 from investigraph.pipeline import run
-from investigraph.settings import DATA_ROOT
 from tests.util import setup_s3_bucket
 
 
@@ -19,8 +21,17 @@ def test_pipeline_local():
     assert len(proxies) == 151
 
 
-# def test_pipeline_local_ray():
-#     os.environ["PREFECT_TASK_RUNNER"] = "ray"
+def test_pipeline_pickle_ctx(gdho):
+    tested = False
+    for source in gdho.extract.sources:
+        ctx = init_context(gdho, source)
+        cloudpickle.dumps(ctx)
+        tested = True
+    assert tested
+
+
+# def test_pipeline_local_ray(monkeypatch):
+#     monkeypatch.setenv("PREFECT_TASK_RUNNER", "ray")
 #     options = FlowOptions(
 #         dataset="eu_authorities", config="./tests/fixtures/eu_authorities.local.yml"
 #     )
@@ -30,11 +41,9 @@ def test_pipeline_local():
 #     proxies = [p for p in smart_read_proxies(out)]
 #     assert len(proxies) == 151
 
-#     del os.environ["TASK_RUNNER"]
 
-
-# def test_pipeline_local_dask():
-#     os.environ["PREFECT_TASK_RUNNER"] = "dask"
+# def test_pipeline_local_dask(monkeypatch):
+#     monkeypatch.setenv("PREFECT_TASK_RUNNER", "dask")
 #     options = FlowOptions(
 #         dataset="eu_authorities", config="./tests/fixtures/eu_authorities.local.yml"
 #     )
@@ -43,8 +52,6 @@ def test_pipeline_local():
 #     out = _run(options)
 #     proxies = [p for p in smart_read_proxies(out)]
 #     assert len(proxies) == 151
-
-#     del os.environ["TASK_RUNNER"]
 
 
 def test_pipeline_from_block(local_block: DatasetBlock):
@@ -60,7 +67,7 @@ def test_pipeline_from_config():
 
 
 def test_pipeline_local_ftmstore():
-    store_uri = f"sqlite:///{DATA_ROOT}/store.db"
+    store_uri = f"sqlite:///{settings.DATA_ROOT}/store.db"
     options = FlowOptions(
         dataset="eu_authorities",
         config="./tests/fixtures/eu_authorities.local.yml",
@@ -90,6 +97,27 @@ def test_pipeline_local_customized():
     options = FlowOptions(
         dataset="eu_authorities", config="./tests/fixtures/eu_authorities.custom.yml"
     )
+    assert run(options)
+
+
+def test_pipeline_chunk_size():
+    options = FlowOptions(
+        dataset="eu_authorities",
+        config="./tests/fixtures/eu_authorities.local.yml",
+        chunk_size=100,
+    )
     out = run(options)
     proxies = [p for p in smart_read_proxies(out)]
     assert len(proxies) == 151
+
+
+def test_pipeline_caching(monkeypatch):
+    monkeypatch.setenv("TASK_CACHE", "true")
+    reload(settings)
+    assert settings.TASK_CACHE is True
+    options = FlowOptions(
+        dataset="eu_authorities", config="./tests/fixtures/eu_authorities.local.yml"
+    )
+    # FIXME this still doesn't work
+    assert run(options)
+    assert run(options)